Fix spurious fontification of "for (; a * b;)" in CC Mode.
This fixes bug #7918 (again).

* cc-engine.el (c-delq-from-dotted-list): New function.
(c-forward-decl-or-cast-1): Return a 4 element list in place of the previous
cons cell - additionally, return a flag indicating whether the declaration
parsed might have been an expression, and the position of the type identifier
in the said declaration.

* cc-fonts.el (c-font-lock-declarations): Remove the code high up in the main
loop that checked every possible declaration position for being a spurious
declaration in a "for" statement.  Replace it by similar code run only when
c-forward-decl-or-cast-1 has indicated it might have parsed an expression.