Fix chaotic indentation of C++ lambda.  Enhance documentation thereof
cc-engine.el (c-looking-at-inexpr-block): qualify an invocation of
c-on-identifier with a check we're not at the _end_ of an identifier.

cc-mode.texi: (Tex title page): Remove @subtitlefont because the perl versions
of texi2dvi haven't implemented it.
(Syntactic Symbols): Note that `inlambda' is also used in C++ Mode, not just
in Pike Mode.
(Statement Block Symbols): Add a section illustrating a C++ lambda function.
(FAQ): Add a question about "excessive" indentation of the contents of a C++
lambda function, and how to get rid of it.